Extract Class Refactoring
moved two methods from the JUnit4TestAdapter class to a new DescriptionEditior class, it will make the classes more cohesive and it will adhere SRP, Tests - ok, Compile - ok
